requirements.txt
python-docx
pathlib2
transformers
torch
scikit-learn
scipy
torchvision
scikit-image